added demand page CRUD-operations

+++ demand/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class DemandPageSalary(models.Model):
     year = models.IntegerField(verbose_name="Год")
-    average_salary = models.FloatField(verbose_name="Средняя зарплата")
+    average_salary = models.IntegerField(verbose_name="Средняя зарплата")
 
 class DemandPageVacanciesCount(models.Model):
     year = models.IntegerField(verbose_name="Год")
